fix user manager token test
